refactor code;
refs fga-eps-mds/2018.2-PDF2CASH#110
