+++ cpp/parser/parser.pro
@@ -15,7 +15,9 @@ DEFINES += QT_DEPRECATED_WARNINGS
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-        main.cpp
+        main.cpp \
+    utils.cpp \
+    Parser.cpp
 
 # Default rules for deployment.
 qnx: target.path = /tmp/$${TARGET}/bin
@@ -34,3 +36,7 @@ else:win32-g++:CONFIG(debug, debug|release): PRE_TARGETDEPS += $$PWD/../poppler-
 else:win32:!win32-g++:CONFIG(release, debug|release): PRE_TARGETDEPS += $$PWD/../poppler-library/build/lib/release/poppler-library.lib
 else:win32:!win32-g++:CONFIG(debug, debug|release): PRE_TARGETDEPS += $$PWD/../poppler-library/build/lib/debug/poppler-library.lib
 else:unix: PRE_TARGETDEPS += $$PWD/../poppler-library/build/lib/libpoppler-library.a
+
+HEADERS += \
+    utils.h \
+    Parser.h
